refactor(ui): WorkspacePanel と TreeEntryNode をモジュール化しファイル行数を削減
