+++ crates/katana-ui/src/views/panels/workspace/mod.rs
@@ -1,5 +1,8 @@
 pub mod breadcrumb;
 pub mod content;
+pub mod dir_node;
+pub mod file_node;
+pub mod header;
 pub mod panel;
 pub mod tree_node;
 
@@ -7,4 +10,4 @@ pub(crate) use breadcrumb::BreadcrumbMenu;
 pub(crate) use panel::WorkspacePanel;
 
 #[cfg(test)]
-pub(crate) use tree_node::FileEntryNode;
+pub(crate) use file_node::FileEntryNode;
